rework this to store the state once
